add createPlayer to Player class
