+++ Player/index.js
@@ -1,3 +1,5 @@
+import { createElem } from "../utils/index.js";
+
 class Player {
     constructor(props) {
         this.player = props.player;
@@ -6,6 +8,7 @@ class Player {
         this.img = props.img;
         this.weapon = [props.weapon];
         this.selector = `player${this.player}`;
+        this.rootSelector = props.rootSelector;
     }
     attack = () => console.log(`${this.name} fight.....`);
     changeHP = (randomVar) => {
@@ -18,11 +21,32 @@ class Player {
     };
 
     elHP = () => {
-        return document.querySelector(`.player${this.player} .life`)
+        return document.querySelector(`.${this.selector} .life`)
     };
 
     renderHP = (getDiv) => {
         return getDiv.style.width = `${this.hp}%`;
+    };
+    createPlayer = () => {
+        const $player = createElem('div', this.selector);
+        const $progressbar = createElem('div', 'progressbar');
+        const $character = createElem('div', 'character');
+        const $life = createElem('div', 'life');
+        const $name = createElem('div', 'name');
+        const $img = createElem('img');
+
+        $life.style.width = `${this.hp}%`;
+        $img.src = this.img;
+        $name.innerText = this.name;
+
+        $player.appendChild($progressbar);
+        $player.appendChild($character);
+        $progressbar.appendChild($life);
+        $progressbar.appendChild($name);
+        $character.appendChild($img);
+        const $root = document.querySelector(`.${this.rootSelector}`)
+        $root.appendChild($player)
+        return $player;
     }
 }
 
@@ -31,11 +55,17 @@ export const player1 = new Player({
     name: 'Scorpion', hp: 100,
     img: 'http://reactmarathon-api.herokuapp.com/assets/scorpion.gif',
     weapon: 'knife',
+    rootSelector: 'arenas'
 })
+
 export const player2 = new Player({
     player: 2,
     name: 'Sub-zero',
     hp: 100,
     img: 'http://reactmarathon-api.herokuapp.com/assets/subzero.gif',
-    weapon: ['ice sword']
+    weapon: ['ice sword'],
+    rootSelector: 'arenas'
 })
+
+
+console.log(player1, player2);
