+++ Utils/index.js
@@ -5,26 +5,26 @@ const createElem = (tag, className) => {
     }
     return $tag;
 }
-const createPlayer = (playerObj) => {
-    const $player = createElem('div', 'player' + playerObj.player);
-    const $progressbar = createElem('div', 'progressbar');
-    const $character = createElem('div', 'character');
-    const $life = createElem('div', 'life');
-    const $name = createElem('div', 'name');
-    const $img = createElem('img');
+// const createPlayer = (playerObj) => {
+//     const $player = createElem('div', 'player' + playerObj.player);
+//     const $progressbar = createElem('div', 'progressbar');
+//     const $character = createElem('div', 'character');
+//     const $life = createElem('div', 'life');
+//     const $name = createElem('div', 'name');
+//     const $img = createElem('img');
 
-    $life.style.width = playerObj.hp + '%';
-    $img.src = playerObj.img;
-    $name.innerText = playerObj.name;
+//     $life.style.width = playerObj.hp + '%';
+//     $img.src = playerObj.img;
+//     $name.innerText = playerObj.name;
 
-    $player.appendChild($progressbar);
-    $player.appendChild($character);
-    $progressbar.appendChild($life);
-    $progressbar.appendChild($name);
-    $character.appendChild($img);
+//     $player.appendChild($progressbar);
+//     $player.appendChild($character);
+//     $progressbar.appendChild($life);
+//     $progressbar.appendChild($name);
+//     $character.appendChild($img);
 
-    return $player;
-}
+//     return $player;
+// }
 
 const getRandom = (numb) => Math.ceil(Math.random() * numb);
 
@@ -51,4 +51,4 @@ const getTime = () => {
 
 }
 
-export { createElem, createPlayer, getRandom, getTime };
+export { createElem, getRandom, getTime };
